Move context to hook folder
