+++ frontend/src/hooks/ConsultantFilterProvider.tsx
@@ -1,16 +1,20 @@
 "use client";
 
 import { Consultant, Department } from "@/types";
-import React, { createContext, ReactNode } from "react";
+import React, { createContext, ReactNode, useState } from "react";
 
 type FilterContextType = {
   consultants: Consultant[];
   departments: Department[];
+  isDisabledHotkeys: Boolean;
+  setIsDisabledHotkeys: React.Dispatch<React.SetStateAction<boolean>>;
 };
 
 export const FilteredContext = createContext<FilterContextType>({
   consultants: [],
   departments: [],
+  isDisabledHotkeys: false,
+  setIsDisabledHotkeys: () => {},
 });
 
 export function ConsultantFilterProvider(props: {
@@ -18,9 +22,16 @@ export function ConsultantFilterProvider(props: {
   departments: Department[];
   children: ReactNode;
 }) {
+  const [isDisabledHotkeys, setIsDisabledHotkeys] = useState(false);
+
   return (
     <FilteredContext.Provider
-      value={{ consultants: props.consultants, departments: props.departments }}
+      value={{
+        consultants: props.consultants,
+        departments: props.departments,
+        isDisabledHotkeys,
+        setIsDisabledHotkeys,
+      }}
     >
       {props.children}
     </FilteredContext.Provider>
